Move demo code, outside of the util package
